Fixed response messages,  updated --help

+++ src/delete_certs.py
@@ -12,24 +12,26 @@ class DeleteCerts:
 
     def delete_certs(self):
         certIds=[]
-        
-        # create instance
-        cert = SslCertificateApiV1.new_instance(
-            crn=self.crn, zone_identifier=self.zone_id, service_name="cis_services")
-        cert.set_service_url(self.endpoint)
-        resp = cert.list_certificates()
-        
-        for cert in resp.result['result']:
-            if self.cis_domain in cert['hosts'] or '*.' + self.cis_domain in cert['hosts']:
-                certIds.append(cert['id']) 
-        
-        #deleting certificates
-        
-        for id in certIds:
+        execute = input("Delete TLS Certificates? Input 'y' to execute: ")
+
+        if execute == 'y':
+            # create instance
             cert = SslCertificateApiV1.new_instance(
                 crn=self.crn, zone_identifier=self.zone_id, service_name="cis_services")
             cert.set_service_url(self.endpoint)
-            resp = cert.delete_certificate(
-                cert_identifier=id)
+            resp = cert.list_certificates()
             
-        print("Deleted",len(certIds),"affiliated certificates")
+            for cert in resp.result['result']:
+                if self.cis_domain in cert['hosts'] or '*.' + self.cis_domain in cert['hosts']:
+                    certIds.append(cert['id']) 
+            
+            #deleting certificates
+            
+            for id in certIds:
+                cert = SslCertificateApiV1.new_instance(
+                    crn=self.crn, zone_identifier=self.zone_id, service_name="cis_services")
+                cert.set_service_url(self.endpoint)
+                resp = cert.delete_certificate(
+                    cert_identifier=id)
+                
+            print("Deleted",len(certIds),"affiliated certificates")
